update name to `config`
